+++ vitest.config.ts
@@ -6,6 +6,7 @@ export default defineConfig({
     environment: "happy-dom",
     coverage: {
       include: ["src"],
+      exclude: ["node_modules/", "dist/", "**/*.d.ts"],
       provider: "v8",
       reporter: ["text", "json", "html"],
     },